// src/utils/msg-types.ts
export enum MsgMouseTypes {
  MouseUp = 'mouse_up',
  MouseDown = 'mouse_down',
  MouseLeft = 'mouse_left',
  MouseRight = 'mouse_right',
  MousePosition = 'mouse_position',
}

export enum MsgDrowingTypes {
    DrawSircle = 'draw_circle',
    DrawRectangle = 'draw_rectangle',
    DrawSquare = 'draw_square',
  }